pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

#
# Project configuration + metadata
#
[project]
name = "objects-api-client-django"
version = "0.3.0"
description = "Easily integrate Objects API in your Django application."
readme = "README.rst"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Maykin", email = "support@maykinmedia.nl"}
]
keywords = ["Objects API", "Client", "Django"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Django",
    "Framework :: Django :: 4.2",
    "Intended Audience :: Developers",
    "Operating System :: Unix",
    "Operating System :: MacOS",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "django >= 4.2",
    "django-solo",
    "requests",
    "zgw_consumers",
]

[project.optional-dependencies]
tests = [
    "requests-mock",
    "pytest",
    "pytest-django",
    "tox",
    "ruff",
]
coverage = ["pytest-cov"]
release = [
    "bump-my-version",
    "twine",
]

[project.urls]
Homepage = "https://github.com/maykinmedia/objects-api-client-django"

#
# Tool configurations
#
[tool.setuptools]
zip-safe = false
include-package-data = true

[tool.setuptools.packages.find]
include = ["objectsapiclient", "objectsapiclient.*"]

[tool.bumpversion]
current_version = "0.3.0"
tag = true
tag_name = "v{new_version}"
commit = true
message = "Bump version to {new_version}"

[[tool.bumpversion.files]]
filename = "pyproject.toml"

[[tool.bumpversion.files]]
filename = "README.rst"
search = ":Version: {current_version}"
replace = ":Version: {new_version}"

[tool.pytest.ini_options]
testpaths = ["tests"]
DJANGO_SETTINGS_MODULE = "testapp.settings"
pythonpath = ["."]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
]

[tool.ruff.lint.isort]
known-first-party = ["objectsapiclient"]
section-order = ["future", "standard-library", "django", "third-party", "first-party", "local-folder"]
combine-as-imports = true

[tool.ruff.lint.isort.sections]
django = ["django"]
